Split oversized runtime modules for Rust gate
