+++ .pm/tasks/task_2412c3dfe33b4b3594273194870eae8f.execution.md
@@ -716,3 +716,35 @@ Example:
   1. Build a fresh Linux package that includes the new non-sequencer execution-binding propagation fix.
   2. Refresh the host-truth deployment stage against that package.
   3. Re-run the destructive validator rebuild on ECS, then verify that storage peer commits are accepted by sequencer and the pair maintains fresh peer heads before moving on to observer seed export.
+
+## 2026-06-07 19:46:05 CST / tpm
+- PR Merge Flow Blocker Found and Fixed:
+  - PR #361 head `4ae0be9e6585c14521b03f0dc79f0f095d483a87` was mergeable, but the `Rust` workflow failed in `required-gate -> Run required test tier`.
+  - Root cause was not behavioral regression; it was the repo file-size gate:
+    - `crates/oasis7/src/bin/oasis7_chain_runtime/governance_registry.rs` at `1804` lines
+    - `crates/oasis7/src/bin/oasis7_chain_runtime/status_payload.rs` at `1274` lines
+    - `crates/oasis7_node/src/node_engine_core.rs` at `1219` lines
+    - `crates/oasis7/src/bin/oasis7_chain_runtime/oasis7_chain_runtime_tests.rs` at `1233` lines
+- Scope-Limited Fix Applied:
+  - Split `governance_registry` inline tests into `crates/oasis7/src/bin/oasis7_chain_runtime/governance_registry_tests.rs`
+  - Split `node_engine_core` inline tests into `crates/oasis7_node/src/node_engine_core_tests.rs`
+  - Split status-payload-oriented runtime tests into `crates/oasis7/src/bin/oasis7_chain_runtime/oasis7_chain_runtime_status_payload_tests.rs`
+  - Extracted observability helper functions from `status_payload.rs` into `crates/oasis7/src/bin/oasis7_chain_runtime/status_payload_observability.rs`
+  - Added the new top-level test module include in `crates/oasis7/src/bin/oasis7_chain_runtime.rs`
+  - No runtime or assertion semantics were intentionally changed; this was a structural shrink to satisfy the repository merge gate while preserving existing coverage.
+- Fresh Verification:
+  - `./scripts/check-rust-file-size.sh`
+  - `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_chain_runtime production_release_policy_status_payload_reports_effective_policy`
+  - `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_chain_runtime network_tier_genesis_bootstrap_initializes_full_governance_world_for_empty_observer`
+  - `./scripts/cargo-dev.sh test -p oasis7_node sequencer_committed_execution_does_not_wait_for_gap_sync`
+  - `git diff --check`
+- Actual Result:
+  - Passed: file-size gate now reports zero oversized code and test files
+  - Passed: moved status payload runtime test
+  - Passed: moved governance bootstrap runtime test
+  - Passed: moved node-engine gap-sync regression test
+  - Passed: diff hygiene
+- Next Action:
+  1. Commit and push the structural shrink fix onto PR #361.
+  2. Re-run or wait for the `Rust` workflow on the new head.
+  3. If checks pass, mark the PR ready and continue normal merge watch.
