+++ crates/oasis7/src/bin/oasis7_chain_runtime.rs
@@ -897,5 +897,8 @@ mod observability_tests;
 #[path = "oasis7_chain_runtime/oasis7_chain_runtime_observability_transport_tests.rs"]
 mod observability_transport_tests;
 #[cfg(test)]
+#[path = "oasis7_chain_runtime/oasis7_chain_runtime_status_payload_tests.rs"]
+mod status_payload_tests;
+#[cfg(test)]
 #[path = "oasis7_chain_runtime/oasis7_chain_runtime_tests.rs"]
 mod tests;
